x64 port
git-svn-id: http://miranda.googlecode.com/svn/trunk@9302 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/src/modules/clist/genmenuopt.cpp
@@ -117,7 +117,7 @@ static int BuildMenuObjectsTree(HWND hwndDlg)
 		return FALSE;
 
 	for ( i=0; i < g_menus.getCount(); i++ ) {
-		if ( g_menus[i]->id == hStatusMenuObject )
+		if ( g_menus[i]->id == (int)hStatusMenuObject )
 			continue;
 
 		tvis.item.lParam  = ( LPARAM )g_menus[i]->id;
